Provide some mounts for customization
The zabbix server container can be customized/extended by providing a
bunch of mounts for things like alertscripts, snmptraps, ssl certs and
more. So adding these mounts from /etc/zabbix/* on the host into the
appropriate directories inside the container so the user can extend
zabbix in the way the container maintainer intended.

Signed-off-by: David Wahlstrom <david.wahlstrom@gmail.com>