Set restart policy and allow ENV pass-in
Setting a restart policy for the containers. This will help provide some
level of recovery from crashes.  Also, allowing ENV to be passed into
the container from a variable, which will allow the user to provide any
ENVs that they want, without us having to define every single possible
ENV var that the container might be able to accept.

Signed-off-by: David Wahlstrom <david.wahlstrom@gmail.com>