Remove additional variable recursive lookup
I wanted to provide a default mysql password, so that things will "just
work" without tinkering too much (knowingly, in an insuceure setup).
However, I was trying to do that with recursive lookups, apparently,
which is a no-no.  So, I created a new "bootstrap" variable called
zabbix_mysql_password which stores the default mysql password. This can
easily be overwritten by the user, but still provides that "just work"
feature I'm looking for.

Signed-off-by: David Wahlstrom <david.wahlstrom@gmail.com>