Add support for proxies
Adding initial support for proxies. At the time being, proxies are
really just servers with a different source container. As such, I copied
the server.yml to a common yml between them and have them both simply
import the tasks from there. However, should things change in the
future, it should be easy to add specific tasks to only the server or
only the proxy with the way I've setup the task import/common yml.

Also, adding support for skipping mysql management. It is concievable
that someone might not want their mysql managed by this module and
instead setup mysql outside of this role. You can already simply change
the server's connection string, so users could already point the server
at an external DB.

Finally, the format of the zabbix variables have changed to a nested
dict. This makes for much easier to read yaml.

Signed-off-by: David Wahlstrom <david.wahlstrom@gmail.com>